requirements.txt
Flask==3.1.1
line_bot_sdk==3.17.1
pandas==2.3.0
python-dotenv==1.1.1
